revset: move resolve_symbol() back to revset module
The only caller is now in `revset.rs`.
